refactor: split FileBrowserDialog into actions hook, preview pane, and action bar
